refactor(client): Split particles.ts into focused modules for LLM optimization
Split the 2061-line particles.ts file into 9 focused modules:

- types.ts: Shared interfaces (Particle, ParticleConfig, DustParticle)
- constants.ts: Animation timing constants
- ParticleEmitter.ts: Core particle emitter class
- outlineTracers.ts: TileOutlineTracer and HexOutlineTracer
- shadows.ts: DropShadow and CircleShadow effects
- dustEffects.ts: DustBurstEffect and MiniDustBurstEffect
- PortalEffect.ts: Hero entrance portal animation
- ParticleManager.ts: Central coordinator for all effects
- index.ts: Barrel file re-exporting public API

Each module is now 100-450 lines, making them easier for LLM agents
to process within context windows. No functional changes.
